we can add float now

+++ 2_calculator.py
@@ -36,8 +36,8 @@ doToday = input("What would you like to do today?")
 #doToday = "add" #HACK
 print("We will " + doToday + " two numbers now.")
 
-var1 = int(input("First number: "))
-var2 = int(input("Second number: "))
+var1 = float(input("First number: "))
+var2 = float(input("Second number: "))
 
 if doToday=="add":
     result = add_two_numbers(var1, var2)
